update sleuthkit to 2.09
This brings mostly stability and security fixes, see sleuthkit changelog for
details

darcs-hash:20070908083231-84eb5-543055b285f3c8e125331ea921592046d0479599.gz

+++ src/filesystems/sleuthkit/fstools/iso9660.c
@@ -1568,7 +1568,7 @@ iso9660_istat(TSK_FS_INFO * fs, FILE * hFile, INUM_T inum,
         OFF_T size = fs_inode->size;
         int rowcount = 0;
 
-        while (size > 0) {
+        while ((int64_t) size > 0) {
             tsk_fprintf(hFile, "%d ", block++);
             size -= fs->block_size;
             rowcount++;
